transfered code to different .cs files
