+++ WebSocketProjects/WebSocketServerWorking/WebSocketServerWorking/WebSocketServerWorking.csproj
@@ -48,6 +48,10 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ConsoleHelpers.cs" />
+    <Compile Include="Main.cs" />
+    <Compile Include="MapData.cs" />
+    <Compile Include="Player.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
